# swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "Optional multiline or single-line description in [CommonMark](http://commonmark.org/help/)\
    \ or HTML."
  version: "0.0.1"
  title: "ProjectAgora"
host: "127.0.0.1:8080"
basePath: "/v1"
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
- "image/png"
- "image/gif"
- "image/jpeg"
paths:
...
  /miniprogram/learner:
    get:
      tags:
      - "Learner"
      - "MiniProgram"
      summary: "小程序端获取成员列表"
      operationId: "miniprogram_learner_get"
      responses:
        200:
          description: "小程序统一使用200状态码，具体信息参照返回码"
          schema:
            type: "object"
            properties: 
              code:
                type: integer
                description: "返回码"
              data:
                description: "数据"
              message:
                description: "信息"
              log:
                description: "错误日志"
              detail:
                description: "更详细的信息"
      x-swagger-router-controller: "swagger_server.controllers.miniprogram_controller"
    post:
      tags:
      - "Learner"
      - "MiniProgram"
      summary: "小程序端注册账户"
      operationId:  "miniprogram_learner_post"
      parameters:
      - name: "learnerPostBody"
        in: "body"
        required: true
        schema:
          type: "object"
          properties:
            sessionKey:
              type: "string"
            encryptedData: 
              type: "string"
            iv:
              type: "string"
            branch:
              type: "string"
            isMentor:
              type: "number"
            givenName:
              type: "string"
            familyName:
              type: "string"
            birthday:
              type: "string"
      responses:
        200:
          description: "小程序统一使用200状态码，具体信息参照返回码"
          schema:
            type: "object"
            properties: 
              code:
                type: integer
                description: "返回码"
              data:
                description: "数据"
              message:
                description: "信息"
              log:
                description: "错误日志"
              detail:
                description: "更详细的信息"
      x-swagger-router-controller: "swagger_server.controllers.miniprogram_controller"
  /miniprogram/booking/{roomCode}:
    get:
      tags:
      - "Booking"
      - "MiniProgram"
      summary: "返回指定房间指定月期的预约信息"
      operationId: "miniprogram_booking_roomCode_get"
      parameters:
      - name: "roomCode"
        in: "path"
        required: true
        type: "string"
      - name: "monthToLoad"
        in: "query"
        required: true
        type: string
      responses:
        200:
          description: "小程序统一使用200状态码，具体信息参照返回码"
          schema:
            type: "object"
            properties: 
              code:
                type: integer
                description: "返回码"
              data:
                description: "数据"
              message:
                description: "信息"
              log:
                description: "错误日志"
              detail:
                description: "更详细的信息"
      x-swagger-router-controller: "swagger_server.controllers.miniprogram_controller"
    post:
      tags:
      - "Booking"
      summary: "在指定房间创建预约信息"
      operationId: "miniprogram_booking_roomCode_post"
      parameters:
      - name: "roomCode"
        in: "path"
        required: true
        type: "string"
      - name: "appointment"
        in: "body"
        required: true
        schema:
          type: object
      responses:
        200:
          description: "小程序统一使用200状态码，具体信息参照返回码"
          schema:
            type: "object"
            properties: 
              code:
                type: integer
                description: "返回码"
              data:
                description: "数据"
              message:
                description: "信息"
              log:
                description: "错误日志"
              detail:
                description: "更详细的信息"
      x-swagger-router-controller: "swagger_server.controllers.miniprogram_controller"
    delete:
      tags:
      - "Booking"
      summary: "在指定房间指定月份删除预约信息"
      operationId: "miniprogram_booking_roomCode_delete"
      parameters:
      - name: "roomCode"
        in: "path"
        required: true
        type: "string"
      - name: "monthToLoad"
        in: "query"
        required: true
        type: string
      - name: "deleteInfo"
        in: "body"
        required: true
        schema:
          type: object
      responses:
        200:
          description: "小程序统一使用200状态码，具体信息参照返回码"
          schema:
            type: "object"
            properties: 
              code:
                type: integer
                description: "返回码"
              data:
                description: "数据"
              message:
                description: "信息"
              log:
                description: "错误日志"
              detail:
                description: "更详细的信息"
      x-swagger-router-controller: "swagger_server.controllers.miniprogram_controller"
...
  /miniprogram/login:
    get:
      tags:
      - "MiniProgram"
      summary: "微信小程序登录 code2session"
      operationId: "miniprogram_login_get"
      parameters:
      - name: "js_code"
        in: "query"
        description: "登录用js_code"
        required: true
        type: "string"
      responses:
        200:
          description: "小程序统一使用200状态码，具体信息参照返回码"
          schema:
            type: "object"
            properties: 
              code:
                type: integer
                description: "返回码"
              data:
                description: "数据"
              message:
                description: "信息"
              log:
                description: "错误日志"
              detail:
                description: "更详细的信息"
      x-swagger-router-controller: "swagger_server.controllers.miniprogram_controller"
    post:
      tags:
      - "MiniProgram"
      summary: "处理登录过程中的授权与鉴权请求"
      operationId: "miniprogram_login_post"
      parameters:
      - in: "body"
        name: "loginPostBody"
        required: true
        schema:
          type: "object"
          properties:
            sessionKey:
              type: "string"
            encryptedData: 
              type: "string"
            iv:
              type: "string"
      responses:
        200:
          description: "小程序统一使用200状态码，具体信息参照返回码"
          schema:
            type: "object"
            properties: 
              code:
                type: integer
                description: "返回码"
              data:
                description: "数据"
              message:
                description: "信息"
              log:
                description: "错误日志"
              detail:
                description: "更详细的信息"
      x-swagger-router-controller: "swagger_server.controllers.miniprogram_controller"
  # /miniprogram_qr/{qrtype}:
  #   get:
  #     tags:
  #     - "MiniProgram"
  #     summary: "处理微信小程序扫码操作"
  #     operationId: "miniprogram_qr_get"
  #     parameters:
  #     - name: "qrtype"
  #       in: "path"
  #       description: "二维码类型"
  #       required: true
  #       type: "string"
  #     responses:
  #       200:
  #         description: "小程序统一使用200状态码，具体信息参照返回码"
  #         schema:
  #           type: "object"
  #           properties: 
  #             code:
  #               type: integer
  #               description: "返回码"
  #             data:
  #               description: "数据"
  #             message:
  #               description: "信息"
  #             log:
  #               description: "错误日志"
  #             detail:
  #               description: "更详细的信息"
  #     x-swagger-router-controller: "swagger_server.controllers.miniprogram_controller"
...
securityDefinitions:
  APIKeyHeader:
    type: "apiKey"
    name: "api_key"
    in: "header"
...